refactor(common-helper): 将 common helper 移动到 utils 目录中
