+++ src/plots/area/adaptor.ts
@@ -1,7 +1,7 @@
 import { Geometry } from '@antv/g2';
 import { deepMix, each } from '@antv/util';
 import { tooltip, interaction, animation, theme } from '../../adaptor/common';
-import { findGeometry } from '../../common/helper';
+import { findGeometry } from '../../utils';
 import { Params } from '../../core/adaptor';
 import { area, point, line } from '../../adaptor/geometries';
 import { flow } from '../../utils';
