+++ src/plots/column/adaptor.ts
@@ -1,6 +1,6 @@
 import { deepMix } from '@antv/util';
 import { Params } from '../../core/adaptor';
-import { findGeometry } from '../../common/helper';
+import { findGeometry } from '../../utils';
 import { tooltip, interaction, animation, theme } from '../../adaptor/common';
 import { interval } from '../../adaptor/geometries';
 import { flow, pick } from '../../utils';
