+++ src/plots/heatmap/adaptor.ts
@@ -1,6 +1,6 @@
 import { deepMix, isFunction, isObject } from '@antv/util';
 import { Params } from '../../core/adaptor';
-import { findGeometry } from '../../common/helper';
+import { findGeometry } from '../../utils';
 import { flow, pick } from '../../utils';
 import { AXIS_META_CONFIG_KEYS, DEFAULT_COLORS } from '../../constant';
 import { tooltip, interaction, animation, theme } from '../../adaptor/common';
