+++ src/plots/line/adaptor.ts
@@ -1,7 +1,7 @@
 import { deepMix } from '@antv/util';
 import { Params } from '../../core/adaptor';
 import { tooltip, interaction, animation, theme } from '../../adaptor/common';
-import { findGeometry } from '../../common/helper';
+import { findGeometry } from '../../utils';
 import { AXIS_META_CONFIG_KEYS } from '../../constant';
 import { point, line } from '../../adaptor/geometries';
 import { flow, pick } from '../../utils';
