+++ src/utils/geometry.ts
