+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export { pick } from './pick';
 export { template } from './template';
 export { log, invariant, LEVEL } from './invariant';
 export { getContainerSize } from './dom';
+export { findGeometry } from './geometry';
